requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.2
django-ckeditor==6.7.2
django-ckeditor-5==0.2.18
django-ckeditor5==0.0.4
django-debug-toolbar==5.2.0
django-js-asset==3.1.2
django-rest-framework==0.1.0
djangorestframework==3.16.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
pipdeptree==2.26.1
platformdirs==4.3.8
pluggy==1.6.0
progress==1.6
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.0
pytest-django==4.11.1
python-dotenv==1.1.0
pytoml==0.1.21
pytz==2025.2
requests==2.32.3
retrying==1.3.4
selenium==4.33.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
textblob==0.17.1
